requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
accelerate>=0.29.0
peft>=0.10.0

# Vision-Language
open-clip-torch>=2.24.0
Pillow>=10.0.0

# Agentic Framework
langgraph>=0.0.55
langchain>=0.1.16
langchain-community>=0.0.36

# RLHF / Alignment
trl>=0.8.6

# Evaluation
ragas>=0.1.7
deepeval>=0.21.0

# MLOps
mlflow>=2.12.0

# Quantization
bitsandbytes>=0.43.0
optimum>=1.19.0

# Utilities
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0.1

# GCP
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.50.0
